feat: implement hierarchical GPU profiler with refined aesthetics and CI fixes
- Added MetricStack for nested profiling and relative percentage logs.
- Implemented rounded timeline bars (SDF) and rounded background box.
- Added colored text shadows and fixed vertical alignment.
- Fixed CI failures by adding test_gpu_profiler to Xvfb runner and using float metrics.
- Refactored input handling to reduce complexity and resolved magic number lint errors.

+++ src/gpu_profiler.c
@@ -129,8 +129,8 @@ void gpu_profiler_begin_frame(GPUProfiler* profiler, uint64_t frame_index)
 
 		adaptive_sampler_add(&profiler->stages[i].sampler,
 		                     (float)duration_ms, read_buf->frame_index);
-		profiler->stages[i].duration_ms = duration_ms;
-		profiler->stages[i].start_offset_ms = offset_ms;
+		profiler->stages[i].duration_ms = (float)duration_ms;
+		profiler->stages[i].start_offset_ms = (float)offset_ms;
 	}
 
 	/* 2. Swap Buffers */
